Add a default token for owner tokens
This is to prevent a token from being sent visibly in the case of not having an owner. We'll replace first name with the from name and email with the from address as it's something to show the end user rather than nothing. Leaving last_name empty so we don't see a full name twice.